+++ src/main/java/gtPlusPlus/xmod/gregtech/common/tileentities/misc/MTESolarHeater.java
@@ -281,7 +281,7 @@ public class MTESolarHeater extends MTETieredMachineBlock {
         if (!hasSolarTower()) {
             this.mTX = aTowerTile.getBaseMetaTileEntity()
                 .getXCoord();
-            this.mTY = (int) aTowerTile.getBaseMetaTileEntity()
+            this.mTY = aTowerTile.getBaseMetaTileEntity()
                 .getYCoord();
             this.mTZ = aTowerTile.getBaseMetaTileEntity()
                 .getZCoord();
